Split weather variable file into different parts for easier maintenance of code
