+++ src/components/3HourForecastData.jsx
@@ -3,18 +3,17 @@ import { useEffect, useState, useMemo, useCallback, memo } from 'react';
 import { FaArrowLeft } from 'react-icons/fa6';
 import { useParams, useNavigate } from 'react-router-dom';
 
+import {
+  TimeZoneShow,
+  SunriseSunsetTimes,
+  VisibilityDesc,
+  WindDirection,
+  WindForce,
+} from '@components/weather/helpers/weatherHelpers';
+import { WeatherIcons, WindArrow } from '@components/weather/WeatherIcons';
 import { useDeviceDetect } from '@hooks/useDeviceDetect';
 import { Footer } from '@utils/footer';
 import { Header } from '@utils/header';
-import {
-  TimeZoneShow,
-  WeatherIcons,
-  WindDirection,
-  VisibilityDesc,
-  WindForce,
-  SunriseSunsetTimes,
-  WindArrow,
-} from '@utils/weatherVariables';
 
 export const ThreeHourForecastData = memo(() => {
   const { lat, lon } = useParams();
