+++ src/components/weather/WeatherDisplay.jsx
@@ -1,432 +1,36 @@
 import axios from 'axios';
 import { useMemo, useCallback, useEffect, useState, memo } from 'react';
-import { AiFillCloud } from 'react-icons/ai'; // Cloudy
 import {
-  BsFillSunFill,
-  BsFillCloudRainHeavyFill,
-  BsFillCloudDrizzleFill,
-  BsFillCloudLightningRainFill,
-  BsFillCloudSnowFill,
-  BsCloudFog,
-  BsFillCloudRainFill,
-  BsFillCloudsFill,
-  BsFillCloudSunFill,
-  BsFillCloudHazeFill,
   BsFillSunriseFill,
   BsFillSunsetFill,
   BsFillCloudSlashFill,
-} from 'react-icons/bs'; // Sunny, Rain, Drizzle, Thunder, Snow, Fog, etc.
-import { FaTemperatureHigh, FaArrowLeft } from 'react-icons/fa'; // Temperature and back arrow icons
-import { GiWindsock } from 'react-icons/gi'; // Wind sock icon
+  BsCloudFog,
+  BsFillCloudRainFill,
+  BsFillCloudDrizzleFill,
+} from 'react-icons/bs';
+import { FaTemperatureHigh, FaArrowLeft } from 'react-icons/fa';
+import { GiWindsock } from 'react-icons/gi';
 import { IoIosArrowUp, IoIosArrowDown } from 'react-icons/io';
-import { LiaLocationArrowSolid } from 'react-icons/lia'; // Wind direction arrow icon
-import { MdModeNight } from 'react-icons/md'; // Night Clear
-import { TbMist, TbWind } from 'react-icons/tb'; // Mist, Windy Clear
-import {
-  WiCloudyWindy,
-  WiDayCloudyWindy,
-  WiDayRainWind,
-  WiDaySnowWind,
-  WiNightFog,
-  WiNightCloudyWindy,
-  WiNightAltPartlyCloudy,
-  WiNightAltCloudy,
-  WiNightAltRainMix,
-  WiNightAltRain,
-  WiNightAltShowers,
-  WiNightAltStormShowers,
-  WiNightAltSnow,
-  WiNightAltRainWind,
-  WiHumidity,
-  WiStrongWind,
-  WiBarometer,
-  WiSmoke,
-  WiDust,
-} from 'react-icons/wi'; // Weather icons for various conditions
+import { WiHumidity, WiStrongWind, WiBarometer } from 'react-icons/wi';
 import { useNavigate, Link } from 'react-router-dom';
 
+import {
+  TimeZoneShow,
+  SunriseSunsetTimes,
+  VisibilityDesc,
+  WindDirection,
+  WindForce,
+  formatTimeDisplay,
+} from '@components/weather/helpers/weatherHelpers';
+import { WeatherIcons, WindArrow } from '@components/weather/WeatherIcons';
 import { Footer } from '@utils/footer';
 import { Header } from '@utils/header';
 
-// Weather icon configuration - declarative approach
-// Each condition defines codes and icons for all 4 states (night/day × windy/calm)
-// Reference: https://openweathermap.org/weather-conditions
-const WEATHER_CONDITIONS = [
-  // Thunderstorm (2xx)
-  {
-    codes: [200, 201, 202, 210, 211, 212, 221, 230, 231, 232],
-    icons: {
-      night: { windy: WiNightAltStormShowers, calm: WiNightAltStormShowers },
-      day: { windy: BsFillCloudLightningRainFill, calm: BsFillCloudLightningRainFill },
-    },
-  },
-  // Light Drizzle (300-302)
-  {
-    codes: [300, 301, 302],
-    icons: {
-      night: { windy: WiNightAltRainWind, calm: WiNightAltShowers },
-      day: { windy: BsFillCloudDrizzleFill, calm: BsFillCloudDrizzleFill },
-    },
-  },
-  // Heavy Drizzle (310-321)
-  {
-    codes: [310, 311, 312, 313, 314, 321],
-    icons: {
-      night: { windy: WiNightAltRainWind, calm: WiNightAltRainMix },
-      day: { windy: BsFillCloudDrizzleFill, calm: BsFillCloudDrizzleFill },
-    },
-  },
-  // Light Rain (500)
-  {
-    codes: [500],
-    icons: {
-      night: { windy: WiNightAltRainWind, calm: WiNightAltRainMix },
-      day: { windy: WiDayRainWind, calm: BsFillCloudRainFill },
-    },
-  },
-  // Heavy Rain (501-531)
-  {
-    codes: [501, 502, 503, 504, 520, 521, 522, 531],
-    icons: {
-      night: { windy: WiNightAltRainWind, calm: WiNightAltRain },
-      day: { windy: BsFillCloudRainHeavyFill, calm: BsFillCloudRainHeavyFill },
-    },
-  },
-  // Freezing Rain (511)
-  {
-    codes: [511],
-    icons: {
-      night: { windy: WiNightAltSnow, calm: WiNightAltSnow },
-      day: { windy: WiDaySnowWind, calm: BsFillCloudSnowFill },
-    },
-  },
-  // Snow (6xx)
-  {
-    codes: [600, 601, 602, 611, 612, 613, 615, 616, 620, 621, 622],
-    icons: {
-      night: { windy: WiNightAltSnow, calm: WiNightAltSnow },
-      day: { windy: WiDaySnowWind, calm: BsFillCloudSnowFill },
-    },
-  },
-  // Mist (701)
-  {
-    codes: [701],
-    icons: {
-      night: { windy: WiNightFog, calm: WiNightFog },
-      day: { windy: TbMist, calm: TbMist },
-    },
-  },
-  // Smoke (711, 762)
-  {
-    codes: [711, 762],
-    icons: {
-      night: { windy: WiSmoke, calm: WiSmoke },
-      day: { windy: WiSmoke, calm: WiSmoke },
-    },
-  },
-  // Haze (721)
-  {
-    codes: [721],
-    icons: {
-      night: { windy: WiNightFog, calm: WiNightFog },
-      day: { windy: BsFillCloudHazeFill, calm: BsFillCloudHazeFill },
-    },
-  },
-  // Dust/Sand (731, 751, 761)
-  {
-    codes: [731, 751, 761],
-    icons: {
-      night: { windy: WiDust, calm: WiDust },
-      day: { windy: WiDust, calm: WiDust },
-    },
-  },
-  // Fog (741)
-  {
-    codes: [741],
-    icons: {
-      night: { windy: WiNightFog, calm: WiNightFog },
-      day: { windy: BsCloudFog, calm: BsCloudFog },
-    },
-  },
-  // Squall (771)
-  {
-    codes: [771],
-    icons: {
-      night: { windy: WiNightCloudyWindy, calm: WiNightAltCloudy },
-      day: { windy: WiCloudyWindy, calm: BsFillCloudsFill },
-    },
-    sizeAdjust: { night: { windy: 20, calm: 20 }, day: { windy: 20, calm: 0 } },
-  },
-  // Tornado (781)
-  {
-    codes: [781],
-    icons: {
-      night: { windy: WiNightCloudyWindy, calm: WiNightAltCloudy },
-      day: { windy: WiCloudyWindy, calm: BsFillCloudsFill },
-    },
-    sizeAdjust: { night: { windy: 20, calm: 20 }, day: { windy: 20, calm: 0 } },
-  },
-  // Clear (800)
-  {
-    codes: [800],
-    icons: {
-      night: { windy: MdModeNight, calm: MdModeNight },
-      day: { windy: TbWind, calm: BsFillSunFill },
-    },
-  },
-  // Few Clouds (801)
-  {
-    codes: [801],
-    icons: {
-      night: { windy: WiNightCloudyWindy, calm: WiNightAltCloudy },
-      day: { windy: WiCloudyWindy, calm: AiFillCloud },
-    },
-    sizeAdjust: { night: { windy: 17, calm: 20 }, day: { windy: 20, calm: 0 } },
-  },
-  // Scattered Clouds (802-803)
-  {
-    codes: [802, 803],
-    icons: {
-      night: { windy: WiNightCloudyWindy, calm: WiNightAltPartlyCloudy },
-      day: { windy: WiDayCloudyWindy, calm: BsFillCloudSunFill },
-    },
-    sizeAdjust: { night: { windy: 20, calm: 30 }, day: { windy: 0, calm: 0 } },
-  },
-  // Overcast Clouds (804)
-  {
-    codes: [804],
-    icons: {
-      night: { windy: WiNightCloudyWindy, calm: WiNightAltCloudy },
-      day: { windy: WiCloudyWindy, calm: BsFillCloudsFill },
-    },
-    sizeAdjust: { night: { windy: 17, calm: 20 }, day: { windy: 20, calm: 0 } },
-  },
-];
-
-// Build the weatherIconsMap from the declarative config
-const buildWeatherIconsMap = () => {
-  const map = {
-    night: { windy: {}, calm: {} },
-    day: { windy: {}, calm: {} },
-  };
-
-  WEATHER_CONDITIONS.forEach(condition => {
-    const { codes, icons, sizeAdjust = {} } = condition;
-
-    codes.forEach(code => {
-      ['night', 'day'].forEach(timeOfDay => {
-        ['windy', 'calm'].forEach(windCondition => {
-          map[timeOfDay][windCondition][code] = {
-            icon: icons[timeOfDay][windCondition],
-            sizeAdjust: sizeAdjust[timeOfDay]?.[windCondition] || 0,
-          };
-        });
-      });
-    });
-  });
-
-  return map;
-};
-
-// Weather icons mapping configuration using OpenWeatherMap weather condition codes
-const weatherIconsMap = buildWeatherIconsMap();
-
-export const WeatherIcons = memo(props => {
-  const currentHourConversion = useMemo(() => {
-    const d = new Date();
-    const baseHour = props.hourConversion || d.getHours();
-    return Math.round((baseHour * 3600 + d.getTimezoneOffset() * 60 + props.timeZone) / 3600);
-  }, [props.timeZone, props.hourConversion]);
-
-  const size = useMemo(
-    () => (props.page === 'single' ? 200 : props.page === 'multiple' ? 50 : 85),
-    [props.page]
-  );
-
-  const isNight = useMemo(
-    () =>
-      currentHourConversion <= props.sunriseHour || currentHourConversion >= props.sunsetHour + 1,
-    [currentHourConversion, props.sunriseHour, props.sunsetHour]
-  );
-
-  const icon = useMemo(() => {
-    const timeOfDay = isNight ? 'night' : 'day';
-    const windCondition = props.windSpeed >= 8.0 ? 'windy' : 'calm';
-
-    // Use weatherId to lookup the icon
-    const iconConfig = weatherIconsMap[timeOfDay]?.[windCondition]?.[props.weatherId];
-
-    if (iconConfig) {
-      const IconComponent = iconConfig.icon;
-      return (
-        <IconComponent size={size + (iconConfig.sizeAdjust || 0)} color={props.color || 'white'} />
-      );
-    }
-
-    // Fallback for any uncovered conditions
-    console.warn(
-      `No weather icon found for: ${timeOfDay}, ${windCondition}, weatherId: ${props.weatherId}`
-    );
-    return <></>;
-  }, [props.weatherId, props.windSpeed, isNight, size, props.color]);
-
-  return icon;
-});
-
-WeatherIcons.displayName = 'WeatherIcons';
-
-export const TimeZoneShow = memo(props => {
-  const formatTimezone = useMemo(() => {
-    // Error handling for invalid inputs
-    if (props.timeZone === undefined || props.timeZone === null) {
-      return 'GMT';
-    }
-
-    const offsetSeconds = parseInt(props.timeZone);
-
-    // Handle invalid parsing or extreme values
-    if (isNaN(offsetSeconds) || Math.abs(offsetSeconds) > 86400) {
-      console.warn(`Invalid timezone offset: ${props.timeZone}`);
-      return 'GMT';
-    }
-
-    // Handle GMT case
-    if (offsetSeconds === 0) {
-      return 'GMT';
-    }
-
-    // Calculate hours and minutes
-    const absOffsetSeconds = Math.abs(offsetSeconds);
-    const hours = Math.floor(absOffsetSeconds / 3600);
-    const minutes = Math.floor((absOffsetSeconds % 3600) / 60);
-
-    // Determine sign
-    const sign = offsetSeconds > 0 ? '+' : '-';
-
-    // Format the timezone string
-    if (minutes === 0) {
-      return `GMT${sign}${hours}`;
-    } else {
-      return `GMT${sign}${hours}:${minutes.toString().padStart(2, '0')}`;
-    }
-  }, [props.timeZone]);
-
-  return formatTimezone || '';
-});
-
-TimeZoneShow.displayName = 'TimeZoneShow';
-
-export const SunriseSunsetTimes = props => {
-  const sunriseTime = new Date(
-    props.sunrise * 1000 + props.timeZone * 1000 + new Date().getTimezoneOffset() * 60 * 1000
-  );
-  const sunsetTime = new Date(
-    props.sunset * 1000 + props.timeZone * 1000 + new Date().getTimezoneOffset() * 60 * 1000
-  );
-
-  return {
-    sunriseHour: sunriseTime.getHours(),
-    sunriseMinute: sunriseTime.getMinutes(),
-    sunsetHour: sunsetTime.getHours(),
-    sunsetMinute: sunsetTime.getMinutes(),
-  };
-};
-
-SunriseSunsetTimes.displayName = 'SunriseSunsetTimes';
-
-export const VisibilityDesc = memo(props => {
-  const visibilityRanges = [
-    { min: 0, max: 50, desc: 'Dense Fog' },
-    { min: 50, max: 200, desc: 'Thick Fog' },
-    { min: 200, max: 500, desc: 'Moderate Fog' },
-    { min: 500, max: 1000, desc: 'Light Fog' },
-    { min: 1000, max: 2000, desc: 'Thin Fog' },
-    { min: 2000, max: 4000, desc: 'Haze' },
-    { min: 4000, max: 10000, desc: 'Light Haze' },
-    { min: 10000, max: 10000, desc: 'Clear' },
-  ];
-
-  const { visibility } = props;
-
-  const visibilityDesc = visibilityRanges.find(
-    range => visibility >= range.min && visibility <= range.max
-  );
-
-  return visibilityDesc ? visibilityDesc.desc : '';
-});
-
-VisibilityDesc.displayName = 'VisibilityDesc';
-
-export const WindDirection = memo(props => {
-  const windDirections = [
-    { min: 348.75, max: 360, dir: 'N' },
-    { min: 0, max: 11.25, dir: 'N' },
-    { min: 11.26, max: 33.75, dir: 'NNE' },
-    { min: 33.75, max: 56.25, dir: 'NE' },
-    { min: 56.25, max: 78.75, dir: 'ENE' },
-    { min: 78.75, max: 101.25, dir: 'E' },
-    { min: 101.25, max: 123.75, dir: 'ESE' },
-    { min: 123.75, max: 146.25, dir: 'SE' },
-    { min: 146.25, max: 168.75, dir: 'SSE' },
-    { min: 168.75, max: 191.25, dir: 'S' },
-    { min: 191.25, max: 213.75, dir: 'SSW' },
-    { min: 213.75, max: 236.25, dir: 'SW' },
-    { min: 236.25, max: 258.75, dir: 'WSW' },
-    { min: 258.75, max: 281.25, dir: 'W' },
-    { min: 281.25, max: 303.75, dir: 'WNW' },
-    { min: 303.75, max: 326.25, dir: 'NW' },
-    { min: 326.25, max: 348.75, dir: 'NNW' },
-  ];
-
-  const { windDegrees } = props;
-
-  const direction = windDirections.find(dir => windDegrees >= dir.min && windDegrees <= dir.max);
-
-  return direction ? direction.dir : '';
-});
-
-WindDirection.displayName = 'WindDirection';
-
-export const WindForce = memo(props => {
-  const windForceRanges = [
-    { min: 0, max: 0.3, force: 'Force 0' },
-    { min: 0.3, max: 1.5, force: 'Force 1' },
-    { min: 1.5, max: 3.3, force: 'Force 2' },
-    { min: 3.3, max: 5.5, force: 'Force 3' },
-    { min: 5.5, max: 8.0, force: 'Force 4' },
-    { min: 8.0, max: 10.8, force: 'Force 5' },
-    { min: 10.8, max: 13.9, force: 'Force 6' },
-    { min: 13.9, max: 17.2, force: 'Force 7' },
-    { min: 17.2, max: 20.7, force: 'Force 8' },
-    { min: 20.7, max: 24.5, force: 'Force 9' },
-    { min: 24.5, max: 28.4, force: 'Force 10' },
-    { min: 28.4, max: 32.6, force: 'Force 11' },
-    { min: 32.6, max: Infinity, force: 'Force 12' },
-  ];
-
-  const { windSpeed } = props;
-
-  const force = windForceRanges.find(range => windSpeed >= range.min && windSpeed < range.max);
-
-  return force ? force.force : '';
-});
-
-WindForce.displayName = 'WindForce';
-
-// WindArrow component for displaying wind direction
-export const WindArrow = memo(({ degrees }) => {
-  return (
-    <LiaLocationArrowSolid
-      className="h-6 w-6 text-red-400"
-      style={{
-        transform: `rotate(${(degrees + 180) % 360}deg)`,
-        transition: 'transform 0.5s ease-in-out',
-      }}
-    />
-  );
-});
-
-WindArrow.displayName = 'WindArrow';
+// Note: Weather-related files have been organized into a better structure:
+// - @config/weatherIconConfig.js: Weather icon configuration and mapping
+// - @utils/weatherHelpers.js: Pure utility functions (timezone, visibility, wind, etc.)
+// - @components/weather/WeatherIcons.jsx: Weather icon components
+// - @components/weather/WeatherDisplay.jsx: Main weather display components (this file)
 
 export const ShowWeather = memo(props => {
   const timeUpdated = useMemo(
@@ -800,11 +404,6 @@ export const ShowWeather = memo(props => {
 });
 
 ShowWeather.displayName = 'ShowWeather';
-
-const formatTimeDisplay = (hour, minute) => {
-  const adjustedHour = hour > 23 ? hour - 24 : hour < 0 ? hour + 24 : hour;
-  return `${adjustedHour.toString().padStart(2, '0')}:${minute.toString().padStart(2, '0')}`;
-};
 
 export const WeatherPopupContent = memo(props => {
   const [currentLocationWeather, setCurrentLocationWeather] = useState(null);
