+++ src/components/weatherSingle.jsx
@@ -3,7 +3,8 @@ import { useState, useEffect, useMemo, useCallback, memo } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import countryList from 'react-select-country-list';
 
-import { ShowWeather, SunriseSunsetTimes } from '@utils/weatherVariables';
+import { SunriseSunsetTimes } from '@components/weather/helpers/weatherHelpers';
+import { ShowWeather } from '@components/weather/WeatherDisplay';
 
 export const GetSingleWeather = memo(() => {
   const { countryCode, city, latitude, longitude } = useParams(); // * Gets city from the url
@@ -130,6 +131,7 @@ export const GetSingleWeather = memo(() => {
       choice={'normal'}
       mainWeather={weather.mainWeather}
       description={weather.description}
+      weatherId={weather.weatherId}
       name={location.name}
       country={location.country}
       temperature={weather.temperature}
